requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
beautifulsoup4
requests
python-dotenv